fraud_detection_app/requirements.txt
Flask==2.0.1
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1
gunicorn==20.1.0
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
dash==2.0.0
plotly==5.3.1
numpy
joblib